asset/requirements.txt
python-bitcoinrpc==1.0
qrcode[PIL]==6.1
